Refactor example population
