+++ bpod_rig/calibration/liquid/utils.py
@@ -2,7 +2,6 @@
 
 import datetime
 import logging
-from pydantic import Field
 import numpy as np
 
 from .models import ValveDataClass, ValveDataManagerClass
@@ -27,41 +26,6 @@ def create_empty_valve_data_manager(
         raise ValueError(f"Unknown source for valve names: {source}")
 
     return valvemanager
-
-
-def add_dummy_measurements(valvemanager: ValveDataManagerClass) -> None:
-    """Add dummy measurements to the valves in manager."""
-    origin_date = datetime.datetime(2000, 1, 1)
-    valve1 = valvemanager.get_valve("Valve1")
-    valve1.add_measurement(22, 2)
-    valve1.add_measurement(66, 9.5)
-    valve1.add_measurement(44, 5)
-    valve1.add_measurement(57, 7.5)
-    valve1.add_measurement(34, 3.5)
-    valve1.lastdatemodified = origin_date
-
-    valve3 = valvemanager.get_valve("Valve3")
-    valve3.add_measurement(22, 1.5)
-    valve3.add_measurement(46, 5.5)
-    valve3.add_measurement(59, 7.5)
-    valve3.add_measurement(35, 3.5)
-    valve3.add_measurement(66, 8.5)
-    valve3.lastdatemodified = origin_date
-
-    valvemanager.metadata.modification_datetime = origin_date
-
-
-def create_default_json() -> str:
-    """Produce a default valve JSON string with 8 valves and dummy measurements."""
-    dummy = create_empty_valve_data_manager()
-    add_dummy_measurements(dummy)
-    jsontext = dummy.to_json(machineid=None)
-    # replace modification datetime with a fixed date for consistency
-    jsontext = jsontext.replace(
-        dummy.metadata.modification_datetime.isoformat(timespec="seconds"),
-        "2000-01-01T00:00:00",
-    )
-    return jsontext
 
 
 def check_valvemanager_user_updated(valvemanager: ValveDataManagerClass) -> bool:
